+++ app/javascript/app/search/filter/search-filters.js
@@ -27,6 +27,57 @@ function init() {
   // Hook reset button on the page and listen for a click event.
   var resetButton = document.getElementById('button-reset');
   resetButton.addEventListener('click', _resetClicked, false);
+
+  var checkboxes = $('#categories input');
+
+  var currentCheckbox;
+  for (var i=0; i < checkboxes.length; i++) {
+    currentCheckbox = checkboxes[i];
+    _checkState('depth',0,currentCheckbox);
+  }
+
+  var curr;
+  for (var l = 0; l < checkboxes.length; l += 1) {
+    curr = checkboxes[l];
+    $(curr).click(_linkClickedHandler);
+  }
+
+  var parentFilterHeaders = document.getElementsByClassName('parent-category-label-container');
+  for (var l = 0; l < parentFilterHeaders.length; l += 1) {
+    curr = parentFilterHeaders[l];
+    curr.addEventListener('click', _handleHeaderClick, false);
+  }
+
+  _openCheckedSections();
+}
+
+function _openCheckedSections() {
+  var checkedBoxes = $('input:checkbox:checked');
+  checkedBoxes.each(function() {
+    if (!($(this).closest('ul').siblings('div').hasClass('selected'))) {
+      _openSection($(this).closest('ul').siblings('div'));
+    }
+  });
+}
+
+function _openSection(element) {
+  var filters = element.nextUntil('depth0');
+  filters.each(function() {
+    $(this).children(0).toggleClass('hide');
+  });
+  $(element).toggleClass('selected');
+  $($(element).find('i')).toggleClass('fa fa-chevron-right fa fa-chevron-down');
+}
+
+function _handleHeaderClick(evt) {
+  _openSection($(evt.currentTarget));
+}
+
+function _linkClickedHandler(evt) {
+  var el = evt.target;
+  if (el.nodeName == 'INPUT') {
+    _checkState('depth', 0, el);
+  }
 }
 
 // The geolocation button was clicked in the location filter.
@@ -41,8 +92,38 @@ function _resetClicked(evt) {
   _location.reset();
   _agency.reset();
 
+  $('input:checkbox:checked').each(function() {
+    $(this).removeAttr('checked');
+  });
+
   evt.preventDefault();
   evt.target.blur();
+}
+
+function _checkState(prefix,depth,checkbox) {
+  var item = $(checkbox).parent(); // parent li item
+  var id = prefix+String(depth);
+  while(!item.hasClass(id)) {
+    depth++;
+    id = prefix + String(depth);
+  }
+
+  id = 'li.' + prefix + String(depth+1);
+  var lnks = $(id, item);
+  var curr;
+  for (var l=0; l < lnks.length; l++) {
+    curr = lnks[l];
+    if (checkbox.checked) {
+      $(curr).removeClass('hide');
+    }
+    else {
+      $(curr).addClass('hide');
+      checkbox = $('input',$(curr));
+      checkbox.prop('checked', false);
+      _checkState(prefix,depth,checkbox);
+    }
+  }
+
 }
 
 export default {